+++ src/const.js
@@ -1,5 +1,13 @@
 const EVENTS_COUNT = 5;
 
+const headerContainer = document.querySelector('.page-header');
+const tripMainContainer = headerContainer.querySelector('.trip-main');
+const tripNavContainer = headerContainer.querySelector('.trip-controls__navigation');
+const tripFiltersContainer = headerContainer.querySelector('.trip-controls__filters');
+
 export {
-  EVENTS_COUNT
+  EVENTS_COUNT,
+  tripMainContainer,
+  tripNavContainer,
+  tripFiltersContainer
 };
